# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/users": {
            "get": {
                "description": "GetAllUsers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "GetAllUsers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名的子串",
                        "name": "username",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取全部用户",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.StatusOKResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            },
            "put": {
                "description": "UserLogin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "UserLogin",
                "parameters": [
                    {
                        "description": "用户名或者邮箱",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正确登陆",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.StatusOKResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数不合法",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusBadRequestResponse"
                        }
                    },
                    "403": {
                        "description": "该用户名或邮箱不存在",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "UserRegister",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "UserRegister",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "邮箱",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusOKResponse"
                        }
                    },
                    "400": {
                        "description": "参数不合法",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusBadRequestResponse"
                        }
                    },
                    "403": {
                        "description": "该邮箱已经被使用",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/avatar": {
            "get": {
                "description": "GetAvatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/jpeg"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "GetAvatar",
                "responses": {
                    "200": {
                        "description": "读取头像成功，data为字节数足",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.StatusOKResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "Accept-Length": {
                                "type": "string",
                                "description": "image's length"
                            },
                            "Content-Disposition": {
                                "type": "string",
                                "description": "attachment; filename=hello.txt"
                            },
                            "Content-Type": {
                                "type": "string",
                                "description": "image/jpeg"
                            }
                        }
                    },
                    "404": {
                        "description": "获取头像失败，下载时出错",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "读取图片失败，处理时出错",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "UploadAvatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "UploadAvatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "用户头像",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传头像成功",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusOKResponse"
                        }
                    },
                    "400": {
                        "description": "请求格式不正确",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusBadRequestResponse"
                        }
                    },
                    "403": {
                        "description": "禁止更改他人资源",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "文件服务错误",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/subscribe": {
            "get": {
                "description": "GetOneUserSubscribe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "GetOneUserSubscribe",
                "responses": {
                    "200": {
                        "description": "获取第{user_id}号用户的关注订阅列表成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.StatusOKResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SubscribeList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "数据库查询出错",
                        "schema": {
                            "$ref": "#/definitions/responses.StatusInternalServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.SubscribeList": {
            "type": "object",
            "properties": {
                "participate_list": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "star_list": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "responses.StatusBadRequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "responses.StatusForbiddenResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 403
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "responses.StatusInternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "responses.StatusOKResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}